pyproject.toml
[sdist]
keep-temp = 1
dist-dir = "dist/source"

[tox]
minversion = "3.4.0"
requires = ["tox-venv", "setuptools >= 30.0.0"]

[tool.poetry]
name = "spp"
version = "0.1.0"
authors = ["Jean-Phillipe Mercier", "Reza Jelveh"]
license = "GPL-3.0"
description = "Python library for processing of seismic data"
keywords = ["seismology", "mining" ]
homepage = "https://api.microquake.org"
repository = "https://git.microquake.org"

packages = [
    { include = "spp", from ="./" },
]

[tool.poetry.scripts]
seismic_platform = 'spp.core.cli:cli'

[tool.poetry.dependencies]
python = "^3.6 || ^3.7"
matplotlib = "*"
obspy = "*"
pymongo = "*"
pyparsing = "*"
scikit-fmm = "*"
prometheus-client = "*"
pyyaml = "*"
toolz = "*"
pytz = "*"
msgpack = "*"
dynaconf = "*"
redis = "*"
tqdm = "*"
tenacity = "*"
apscheduler = "*"
scipy = "^1.2"
pathos = "^0.2.3"
numpy = "^1.16"
h5py = "^2.9"
pandas = "^0.24.2"
loguru = "^0.2.5"
grpcio-tools = "*"
click = "^7.0"
microquake-hashwrap = "^0.0.4"
confluent-kafka = "^1.0"
keras = "^2.2"
tensorflow = "^1.13"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
autopep8 = "*"
ipython = "^7.5"
pytest-parallel = "^0.0.9"
